removed replay checkbox, mean only for MiRa6024 and SY6011

+++ radar_eval/communication/radar_bridge_usb_device.py
@@ -216,7 +216,7 @@ class MIRA_USB_SPI_BRIDGE():
 
         USB_SPI_BRIDGE_DATA_ALLOCATION = np.uint32(self.config.get("MIRA_USB_SPI_BRIDGE",
                                                                    f"USB_SPI_BRIDGE_DATA_ALLOCATION_{self.radar_param.mon.sykno_product_name}"))
-        DATA_ALLOCATION = np.uint32(USB_SPI_BRIDGE_DATA_ALLOCATION+self.radar_param.sys.n_fifo_overhead*9)
+        DATA_ALLOCATION = np.uint32(np.uint32(USB_SPI_BRIDGE_DATA_ALLOCATION) + np.uint32(self.radar_param.sys.n_fifo_overhead) * 9)
 
         raw_data = np.zeros(1,)
         while not process_stop_event.is_set():
